+++ src/index.tsx
@@ -2,4 +2,4 @@ export * from './base_sdk';
 export * from './models/enums';
 export * from './models/errors';
 export * from './models/objects';
-export * from './features/auth';
+export * from './managers/auth';
