+++ src/managers/auth.ts
