+++ src/managers/payment.ts
