+++ src/managers/reader.ts
