+++ src/managers/settings.ts
